+++ backend/core/js/backend.js
@@ -22,13 +22,10 @@ jsBackend =
 	// init, something like a constructor
 	init: function()
 	{
-		// get url and split into chunks
-		var chunks = document.location.pathname.split('/');
-
 		// set some properties
-		jsBackend.current.module = chunks[3];
-		jsBackend.current.action = chunks[4];
-		jsBackend.current.language = chunks[2];
+		jsBackend.current.module = '{$MODULE}';
+		jsBackend.current.action = '{$ACTION}';
+		jsBackend.current.language = '{$LANGUAGE}';
 
 		// init stuff
 		jsBackend.initAjax();
@@ -1156,14 +1153,10 @@ jsBackend.forms =
 		});
 
 		// not changed?
-		if(!changed) {
-			// prevent default
-			/*
-			 * I know this line triggers errors, if you remove it the unload won't work anymore.
-			 * Probably you'll fix this by passing the evt as an argument of the function, well this will break the functionality also..
-			 * Uhu, a "wtf" is in place..
-			 */
-			evt.preventDefault();
+		if(!changed)
+		{
+			// prevent default action from being executed
+			if(evt) evt.preventDefault();
 
 			// unbind the event
 			$(window).unbind('beforeunload');