+++ backend/modules/dashboard/js/dashboard.js
@@ -145,7 +145,7 @@ jsBackend.dashboard =
 
 		// hide removed
 		$('.sortableWidget.isRemoved').hide();
-		
+
 		// make the call
 		$.ajax(
 		{